% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb.R
\name{as_flextable.rbind_tb}
\alias{as_flextable.rbind_tb}
\title{Convert rbind_tb Object to Flextable}
\usage{
as_flextable.rbind_tb(x, ...)
}
\arguments{
\item{x}{An \code{rbind_tb} object.}

\item{...}{Additional arguments passed to \code{flextable::flextable()}.}
}
\value{
A \code{flextable} object.
}
\description{
Convert rbind_tb Object to Flextable
}
