% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Package Initialization Hooks}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Library name path.}

\item{pkgname}{Package name string.}
}
\description{
Package Initialization Hooks
}
\keyword{internal}
