% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb.R
\name{.resolve_label}
\alias{.resolve_label}
\title{Helper function for internal label resolution}
\usage{
.resolve_label(var_name, data, labels_arg)
}
\description{
Helper function for internal label resolution
}
\keyword{internal}
