% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb.R
\name{print.rbind_tb}
\alias{print.rbind_tb}
\title{Print Method for rbind_tb Objects}
\usage{
\method{print}{rbind_tb}(x, digits = NULL, ...)
}
\arguments{
\item{x}{An \code{rbind_tb} object.}

\item{digits}{Minimum number of significant digits to be printed.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for rbind_tb Objects
}
