% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb.R
\name{rbind}
\alias{rbind}
\title{Combine Objects by Rows}
\usage{
rbind(..., deparse.level = 1)
}
\arguments{
\item{...}{Objects to be combined.}

\item{deparse.level}{Integer controlling label deparsing.}
}
\value{
A combined object.
}
\description{
Combine Objects by Rows
}
