% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb.R
\name{rbind.tb}
\alias{rbind.tb}
\title{Combine tb Objects by Rows}
\usage{
rbind.tb(..., deparse.level = 1)
}
\arguments{
\item{...}{Objects of class \code{tb} to be combined.}

\item{deparse.level}{Integer controlling label deparsing (unused).}
}
\value{
A combined object of class \code{c("rbind_tb", "simtab")}.
}
\description{
Vertical stacking of \code{tb} objects to create multi-variable tables (foundations for Table 1).
}
