% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb.R
\name{simtab_theme}
\alias{simtab_theme}
\title{SimtablR House Style Theme for Flextable}
\usage{
simtab_theme(ft)
}
\arguments{
\item{ft}{A flextable object.}
}
\value{
A flextable object.
}
\description{
SimtablR House Style Theme for Flextable
}
